Add file separeted to computed fine movement features.
