+++ app/src/main/res/values/strings.xml
@@ -157,6 +157,10 @@
     <string name="intelligibility">Intelligibility</string>
     <string name="patient">Patient</string>
     <string name="control">Control</string>
+    <string name="tapPosition">Error Position</string>
+    <string name="tapHits">% Hits Ladybug</string>
+    <string name="barHits">% Hits Sliding</string>
+    <string name="tapVelocity">Tapping Velocity</string>
 
 
 </resources>
